Fix broken config items. Randomize stone bricks in dynamically generated legs.

+++ src/main/java/com/yungnickyoung/minecraft/betterstrongholds/config/ConfigGeneral.java
@@ -7,10 +7,10 @@ public class ConfigGeneral {
     public final ForgeConfigSpec.ConfigValue<Integer> strongholdStartMaxY;
     public final ForgeConfigSpec.ConfigValue<Integer> strongholdMaxY;
     public final ForgeConfigSpec.ConfigValue<Integer> strongholdSize;
-    public final ForgeConfigSpec.ConfigValue<Float> cobwebReplacementChanceNormal;
-    public final ForgeConfigSpec.ConfigValue<Float> cobwebReplacementChanceSpawner;
-    public final ForgeConfigSpec.ConfigValue<Float> torchSpawnRate;
-    public final ForgeConfigSpec.ConfigValue<Float> lanternSpawnRate;
+    public final ForgeConfigSpec.ConfigValue<Double> cobwebReplacementChanceNormal;
+    public final ForgeConfigSpec.ConfigValue<Double> cobwebReplacementChanceSpawner;
+    public final ForgeConfigSpec.ConfigValue<Double> torchSpawnRate;
+    public final ForgeConfigSpec.ConfigValue<Double> lanternSpawnRate;
 
     public ConfigGeneral(final ForgeConfigSpec.Builder BUILDER) {
         BUILDER
@@ -60,28 +60,28 @@ public class ConfigGeneral {
                 " The rate at which cobwebs will spawn in various parts of the stronghold.\n" +
                 " Default: 0.1")
             .worldRestart()
-            .define("Cobweb Spawn Rate (NORMAL)", 0.1f);
+            .define("Cobweb Spawn Rate (NORMAL)", 0.1);
 
         cobwebReplacementChanceSpawner = BUILDER
             .comment(
                 " The rate at which cobwebs will spawn around spider spawners in libraries.\n" +
                 " Default: 0.3")
             .worldRestart()
-            .define("Cobweb Spawn Rate (SPAWNER)", 0.3f);
+            .define("Cobweb Spawn Rate (SPAWNER)", 0.3);
 
         torchSpawnRate = BUILDER
             .comment(
                 " The rate at which torches spawn throughout the stronghold.\n" +
                 " Default: 0.1")
             .worldRestart()
-            .define("Torch Spawn Rate", 0.1f);
+            .define("Torch Spawn Rate", 0.1);
 
         lanternSpawnRate = BUILDER
             .comment(
                 " The rate at which lanterns spawn throughout the stronghold.\n" +
                 " Default: 0.2")
             .worldRestart()
-            .define("Lantern Spawn Rate", 0.2f);
+            .define("Lantern Spawn Rate", 0.2);
 
         BUILDER.pop();
     }
